+++ code/classeine-lib/clsn/ui/defaults/DefaultDarkUITheme.cpp
@@ -4,20 +4,48 @@
 
 namespace clsn::ui::defaults
 {
-    void DefaultDarkUITheme::populateDefaults()
+    auto DefaultDarkUITheme::getBevelUpColor() const -> Color
     {
-        add("", "bevelUpColor", Color{0, 92, 155 });
-        add("", "bevelDownColor", Color{0, 49, 83 });
-        add("", "containerBackgroundColor", Color{0, 25, 42});
-        add("", "containerForegroundColor", Colors::makeWhite());
-        add("", "controlBackgroundColor", Color{0, 49, 83});
-        add("", "controlForegroundColor", Color{240, 240, 250});
-        add("", "disabledBackgroundColor", Color{0x60, 0x60, 0x60});
-        add("", "disabledForegroundColor", Color{0, 25, 42});
+        return Color{0, 92, 155 };
+    }
 
-        add("", "font", Font{"Nimbus", FontStyle::REGULAR, 14});
+    auto DefaultDarkUITheme::getBevelDownColor() const -> Color
+    {
+        return Color{0, 49, 83 };
+    }
 
-        add("Button", "pressedBackgroundColor", Color{0, 35, 60});
-        add("Button", "preferredSize", Dimension{96, 32});
+    auto DefaultDarkUITheme::getContainerBackgroundColor() const -> Color
+    {
+        return Color{0, 25, 42};
+    }
+
+    auto DefaultDarkUITheme::getContainerForegroundColor() const -> Color
+    {
+        return Colors::makeWhite();
+    }
+
+    auto DefaultDarkUITheme::getControlBackgroundColor() const -> Color
+    {
+        return Color{0, 49, 83};
+    }
+
+    auto DefaultDarkUITheme::getControlForegroundColor() const -> Color
+    {
+        return Color{240, 240, 250};
+    }
+
+    auto DefaultDarkUITheme::getDisabledBackgroundColor() const -> Color
+    {
+        return Color{0x60, 0x60, 0x60};
+    }
+
+    auto DefaultDarkUITheme::getDisabledForegroundColor() const -> Color
+    {
+        return Color{0, 25, 42};
+    }
+
+    auto DefaultDarkUITheme::getButtonPressedBackgroundColor() const -> Color
+    {
+        return Color{0, 35, 60};
     }
 }