+++ code/classeine-lib/clsn/ui/defaults/DefaultLightUITheme.cpp
@@ -4,20 +4,49 @@
 
 namespace clsn::ui::defaults
 {
-    void DefaultLightUITheme::populateDefaults()
+    auto DefaultLightUITheme::getBevelUpColor() const -> Color
     {
-        add("", "bevelUpColor", Color{0, 92, 155 });
-        add("", "bevelDownColor", Color{0, 49, 83 });
-        add("", "containerBackgroundColor", Colors::makeWhite());
-        add("", "containerForegroundColor", Colors::makeBlack());
-        add("", "controlBackgroundColor", Color{0xE0, 0xE0, 0xE0});
-        add("", "controlForegroundColor", Color{0, 49, 83});
-        add("", "disabledBackgroundColor", Color{0x60, 0x60, 0x60});
-        add("", "disabledForegroundColor", Color{160, 160, 160});
-
-        add("", "font", Font{"Nimbus", FontStyle::REGULAR, 14});
-
-        add("Button", "pressedBackgroundColor", Color{192, 192, 192});
-        add("Button", "preferredSize", Dimension{96, 32});
+        return Color{0, 92, 155 };
     }
+
+    auto DefaultLightUITheme::getBevelDownColor() const -> Color
+    {
+        return Color{0, 49, 83 };
+    }
+
+    auto DefaultLightUITheme::getContainerBackgroundColor() const -> Color
+    {
+        return Colors::makeWhite();
+    }
+
+    auto DefaultLightUITheme::getContainerForegroundColor() const -> Color
+    {
+        return Colors::makeBlack();
+    }
+
+    auto DefaultLightUITheme::getControlBackgroundColor() const -> Color
+    {
+        return Color{0xE0, 0xE0, 0xE0};
+    }
+
+    auto DefaultLightUITheme::getControlForegroundColor() const -> Color
+    {
+        return Color{0, 49, 83};
+    }
+
+    auto DefaultLightUITheme::getDisabledBackgroundColor() const -> Color
+    {
+        return Color{0x60, 0x60, 0x60};
+    }
+
+    auto DefaultLightUITheme::getDisabledForegroundColor() const -> Color
+    {
+        ;return Color{160, 160, 160};
+    }
+
+    auto DefaultLightUITheme::getButtonPressedBackgroundColor() const -> Color
+    {
+        return Color{192, 192, 192};
+    }
+
 }